refactor: remove deprecated analysis classes and services for improved code clarity
- Deleted multiple analysis classes including ActionStatsAnalyzer, AgentAnalyzer, BehaviorClusteringAnalyzer, CausalAnalyzer, DecisionPatternAnalyzer, ResourceImpactAnalyzer, SequencePatternAnalyzer, SimulationAnalyzer, TemporalPatternAnalyzer, and their associated methods to streamline the analysis framework.
- Removed the ActionsService and related service files, consolidating functionality under a more unified structure.
- Updated import paths in test files to reflect the removal of deprecated classes and services, ensuring consistency across the codebase.
- Enhanced overall code maintainability by eliminating obsolete components and focusing on a more cohesive architecture for agent action analysis.

These changes significantly improve the clarity and organization of the codebase, facilitating easier maintenance and future development.
