+++ database/analyzers/agent_analyzer.py
