+++ database/analyzers/resource_impact_analyzer.py
