+++ database/analyzers/sequence_pattern_analyzer.py
