+++ database/analyzers/simulation_analyzer.py
