+++ database/analyzers/temporal_pattern_analyzer.py
